Changes to be committed: 	new file:   .vscode/settings.json 	modified:   include/menu.h 	modified:   platformio.ini 	modified:   src/main.cpp 	modified:   src/menus.cpp
added new menu pages in .h and menus cpp

this includes a starting page which the
device will show on startup, this includes a
readout of effects as well as level bars and bypass
indicators.

started working on bars functionality varying
width using analogue pot input, currently this
has a weird jumpy behaviour in the middle of the
range, see comment for more information.

added variables for all bar fill widths only one
is used currently.