Changes to be committed: 	modified:   src/menus.cpp
added new serial debugging to button control function

row and col counters are working as expected

added extra delay to de bounce buttons, although
the display is kinda fucky there are discrete changes
to the UI when buttons are pressed.

TODO: although it wont work perfectly, it is time to
add the code for the other highlighted elements as currently
only the rectangle aroud the left text collumn is coded
to work.

maybe refer back to the array structure

also add some serial debugging features to the

loc x and loc y vars to check they arent getting strange
values

if theyre fine then it may be the display timing being
weird wit strange use of delays which the program
currently relies upon. Really need to change this.